feat: Adds Doubao configuration and integration for enhanced AI service interaction 增加豆包大模型配置end-point(接入点) 优化代码结构，解耦业务逻辑和与豆包交互逻辑
