+++ paicoding-service/src/main/java/com/github/paicoding/forum/service/chatai/service/impl/doubao/DoubaoConfig.java
@@ -1,0 +1,21 @@
+package com.github.paicoding.forum.service.chatai.service.impl.doubao;
+
+import lombok.Data;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+import javax.annotation.PostConstruct;
+
+@Data
+@Configuration
+@ConfigurationProperties(prefix = "doubao")
+public class DoubaoConfig{
+
+    @Value("${doubao.api-key}")
+    private String apiKey;
+    @Value("${doubao.api-host}")
+    private String apiHost;
+    @Value("${doubao.end-point}")
+    private String endPoint;
+}
